Move into definitions.xml
